+++ src/graphql/root-schema.js
@@ -2,17 +2,21 @@ const { gql } = require('apollo-server')
 
 const { context, onConnect } = require('./context')
 
+const { scalarDefs, scalarResolvers } = require('./scalars')
+
 const authSchema = require('./schema/auth')
 const userSchema = require('./schema/user')
 const deviceSchema = require('./schema/device')
+const measurementSchema = require('./schema/measurement')
 
 const rootTypeDefs = gql`
 	type RootSubsription {
 		user(email: String!): User!
+		newMeasurements(mac: String!): Measurement!
 	}
 
 	type RootQuery {
-		user(email: String): User @requiresAuth(rolesAccepted: [USER, DEVICE])
+		user(email: String): User
 		device(mac: String): Device
 		loginUser(email: String!, password: String!): UserAuthData!
 		loginDevice(mac: String!, pin: Int!): DeviceAuthData!
@@ -34,6 +38,12 @@ const rootTypeDefs = gql`
 
 		txBeacon: String! @requiresAuth(rolesAccepted: [DEVICE])
 
+		txMeasurement(
+			type: MeasurementType!
+			environment: MeasurementEnvironment
+			value: Float!
+		): Measurement! @requiresAuth(rolesAccepted: [DEVICE])
+
 		createDevice(deviceInput: DeviceInput!): Device!
 			@requiresAuth(rolesAccepted: [DEPLOYER, ADMIN])
 
@@ -49,28 +59,35 @@ const rootTypeDefs = gql`
 
 const rootSchema = {
 	typeDefs: [
+		scalarDefs,
 		authSchema.typeDefs,
 		userSchema.typeDefs,
 		deviceSchema.typeDefs,
+		measurementSchema.typeDefs,
 		rootTypeDefs,
 	],
 	resolvers: {
+		...scalarResolvers,
 		RootSubsription: {
 			...userSchema.subscriptionResolvers,
+			...measurementSchema.subscriptionResolvers,
 		},
 		RootQuery: {
 			...authSchema.queryResolvers,
 			...userSchema.queryResolvers,
 			...deviceSchema.queryResolvers,
+			...measurementSchema.queryResolvers,
 		},
 		RootMutation: {
 			...authSchema.mutationResolvers,
 			...userSchema.mutationResolvers,
 			...deviceSchema.mutationResolvers,
+			...measurementSchema.mutationResolvers,
 		},
 		AuthData: authSchema.AuthDataResolver,
 		User: userSchema.UserResolver,
 		Device: deviceSchema.DeviceResolver,
+		Measurement: measurementSchema.MeasurementResolver,
 	},
 	context,
 	subscriptions: { onConnect },