+++ src/models/user.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose')
 const bcryptPlugin = require('mongoose-bcrypt')
 
-const Schema = mongoose.Schema
+const { Schema } = mongoose
 
 const userSchema = new Schema(
 	{